requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
numpy
Pillow
torch
torchvision
ultralytics
lapx>=0.5.2
paddlepaddle
paddleocr
python-multipart